Free up regex ANYOF bit.
This uses a regnode type, of which we have many available, to free up
a bit in the ANYOF regnode flag field, of which we have none, and are
trying to have the same bit do double duty.  This will enable us to
remove some of that double duty in the next commit.